Made the visited types a member of the mapper
This allows avoiding double checks for types in different calls in
GenericArgumentBinderContext.TryBind method. It is ok since the generic
types are shared among the parameters of the method, and the mapper is
created per call.
This also fixes a logical flaw from a previous commit where a pair of
parameterType/realType is considered ok if revisited - even if that
visit resulted with a mismatch.